Add daily project update email
This commit adds a daily email that features a list of added, removed,
and changed projects. Hopefully this architecture can be extended to
other objects as well.

To track the additions/removals/changes, I've pulled in a pub/sub
library called Wisper. By connecting it to ActiveRecord callbacks, we
can decouple the responsibilities into a pattern that is quite
convenient for batching up a bunch of events into an email like this.